Refactor: extract ctaphid interface
