+++ src/authenticator.rs
@@ -2,11 +2,6 @@ use core::borrow::Borrow;
 use core::convert::TryInto;
 use core::time::Duration;
 
-#[cfg(feature = "ctaphid")]
-use ctaphid_dispatch::app::{self as hid, Command as HidCommand, Message};
-#[cfg(feature = "ctaphid")]
-use ctaphid_dispatch::command::VendorCommand;
-
 use flexiber::{Encodable, EncodableHeapless};
 use iso7816::{Data, Status};
 use trussed::{
@@ -367,7 +362,6 @@ where
 
         let mut file_index = 0;
         while let Some(credential) = maybe_credential {
-
             // keep track, in case we need continuation
             file_index += 1;
             self.state.runtime.previously = Some(CommandState::ListCredentials(file_index));
@@ -935,12 +929,11 @@ where
     }
 
     fn delay_on_failure(&mut self) {
-        
+
         // TODO block for the time defined in the constant
         // DESIGN allow only a couple of failures per power cycle? Similarly to the FIDO2 PIN
     }
 }
-
 
 impl<T> iso7816::App for Authenticator<T> {
     fn aid(&self) -> iso7816::Aid {
@@ -971,56 +964,5 @@ where
         reply: &mut Data<R>,
     ) -> Result {
         self.respond(apdu, reply)
-    }
-}
-
-#[cfg(feature = "ctaphid")]
-const OTP_CCID: VendorCommand = VendorCommand::H70;
-
-#[cfg(feature = "ctaphid")]
-impl<T> hid::App for Authenticator<T>
-where
-    T: client::Client
-        + client::HmacSha1
-        + client::HmacSha256
-        + client::Sha256
-        + client::Chacha8Poly1305,
-{
-    fn commands(&self) -> &'static [HidCommand] {
-        &[HidCommand::Vendor(OTP_CCID)]
-    }
-
-    fn call(
-        &mut self,
-        command: HidCommand,
-        input_data: &Message,
-        response: &mut Message,
-    ) -> hid::AppResult {
-        const MAX_COMMAND_LENGTH: usize = 255;
-        match command {
-            HidCommand::Vendor(OTP_CCID) => {
-                let arr: [u8; 2] = Status::Success.into();
-                response.extend(arr);
-                let ctap_to_iso7816_command =
-                    iso7816::Command::<MAX_COMMAND_LENGTH>::try_from(input_data).map_err(|_e| {
-                        response.clear();
-                        debug_now!("ISO conversion error: {:?}", _e);
-                        hid::Error::InvalidLength
-                    })?;
-                self.respond(&ctap_to_iso7816_command, response)
-                    .map_err(|e| {
-                        debug_now!("OTP command execution error: {:?}", e);
-                        let arr: [u8; 2] = e.into();
-                        response.clear();
-                        response.extend(arr);
-                        hid::Error::InvalidCommand
-                    })
-                    .ok();
-            }
-            _ => {
-                return Err(hid::Error::InvalidCommand);
-            }
-        }
-        Ok(())
     }
 }
