+++ src/lib.rs
@@ -13,7 +13,8 @@ pub use authenticator::Authenticator;
 pub mod calculate;
 pub mod command;
 pub use command::Command;
-mod credential;
+pub mod credential;
+pub mod ctaphid;
 pub mod encrypted_container;
 pub mod oath;
 pub mod state;
